django_blog/mysite/requirement.txt
awesome-slugify==1.6.5
Django==1.10.1
django-extensions==2.0.7
django-shell-plus==1.1.7
dwebsocket==0.4.2
et-xmlfile==1.0.1
ipython==3.0.0
jdcal==1.4
mysqlclient==1.3.12
openpyxl==2.5.3
Pillow==5.1.0
PyMySQL==0.8.0
pytz==2018.4
redis==2.10.6
regex==2018.6.9
six==1.11.0
typing==3.6.4
Unidecode==0.4.21
Markdown-2.6.11
Pillow==5.1.0  # 图片处理
sorl-thumbnail  # 缩略图
django-braces  # django开发中常用的mix-in类